compile-time enforcement of printf strings

+++ firmware/controllers/trigger/decoders/trigger_structure.cpp
@@ -709,7 +709,7 @@ void TriggerWaveform::initializeTriggerWaveform(operation_mode_e triggerOperatio
 
 	default:
 		setShapeDefinitionError(true);
-		warning(ObdCode::CUSTOM_ERR_NO_SHAPE, "initializeTriggerWaveform() not implemented: %d", triggerConfig.TriggerType.type);
+		warning(ObdCode::CUSTOM_ERR_NO_SHAPE, "initializeTriggerWaveform() not implemented: %d", (int)triggerConfig.TriggerType.type);
 	}
 
 	/**
